src_fastapi/requirements.txt
fastapi[all]==0.104.1
uvicorn==0.24.0.post1
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# test
pytest==7.4.3
pytest-env==1.1.3
pytest-mock==3.12.0
factory-boy==3.2.1

debugpy
